move and test getConvertedPredictedLabel
